Website structure and future development planning
This first addition to this repository for my personal website contains the initial version of my portfolio website which is based on some wireframes I did when analysing Spotify, and recreating it my way. I've laid the foundation for more exciting features to be added in the upcoming versions as well as figure out how to embed my documentation on a webpage.